converted to f64

+++ src/common.rs
@@ -1,7 +1,7 @@
 use image::RgbImage;
 use rand::Rng;
 use serde::{Deserialize, Serialize};
-use std::f32::consts::PI;
+use std::f64::consts::PI;
 
 use crate::camera::*;
 use crate::geometry::*;
@@ -20,7 +20,7 @@ pub fn reflect(incident: Vec3f, normal: Vec3f) -> Vec3f {
 
 // https://www.scratchapixel.com/lessons/3d-basic-rendering/introduction-to-shading/reflection-refraction-fresnel.html
 #[allow(dead_code)]
-pub fn refract(incident: Vec3f, normal: Vec3f, ior: f32) -> Vec3f {
+pub fn refract(incident: Vec3f, normal: Vec3f, ior: f64) -> Vec3f {
     let mut cosi = Vec3f::dot(incident, normal);
     let mut etai = 1.0;
     let mut etat = ior;
@@ -46,7 +46,7 @@ pub fn refract(incident: Vec3f, normal: Vec3f, ior: f32) -> Vec3f {
 
 // https://registry.khronos.org/OpenGL-Refpages/gl4/html/refract.xhtml
 #[allow(dead_code)]
-pub fn refract_glsl(incident: Vec3f, normal: Vec3f, eta: f32) -> Vec3f {
+pub fn refract_glsl(incident: Vec3f, normal: Vec3f, eta: f64) -> Vec3f {
     let cos_incident = Vec3::dot(normal, incident);
     let k = 1.0 - eta * eta * (1.0 - cos_incident * cos_incident);
     if k < 0.0 {
@@ -58,18 +58,18 @@ pub fn refract_glsl(incident: Vec3f, normal: Vec3f, eta: f32) -> Vec3f {
 }
 
 ///
-pub fn fresnel(incident: Vec3f, normal: Vec3f, ior: f32) -> f32 {
+pub fn fresnel(incident: Vec3f, normal: Vec3f, ior: f64) -> f64 {
     let mut cosi = Vec3f::dot(incident, normal);
     let etai = 1.0;
     let etat = ior;
 
-    let sint = etai / etat * f32::sqrt(f32::max(0.0, 1.0 - cosi * cosi));
+    let sint = etai / etat * f64::sqrt(f64::max(0.0, 1.0 - cosi * cosi));
 
     let kr = if sint >= 1.0 {
         // Total internal reflection
         1.0
     } else {
-        let cost = f32::sqrt(f32::max(0.0, 1.0 - sint * sint));
+        let cost = f64::sqrt(f64::max(0.0, 1.0 - sint * sint));
         cosi = cosi.abs();
         let rs = ((etat * cosi) - (etai * cost)) / ((etat * cosi) + (etai * cost));
         let rp = ((etai * cosi) - (etat * cost)) / ((etai * cosi) + (etat * cost));
@@ -81,7 +81,7 @@ pub fn fresnel(incident: Vec3f, normal: Vec3f, ior: f32) -> f32 {
 
 /// Returns vector based on spherical coordinates
 /// But: in our coordinate system, y is up
-pub fn from_spherical(theta: f32, phi: f32) -> Vec3f {
+pub fn from_spherical(theta: f64, phi: f64) -> Vec3f {
     let sin_phi = phi.sin();
     let cos_phi = phi.cos();
     let sin_theta = theta.sin();
@@ -95,7 +95,7 @@ pub fn uniform_hemisphere() -> Vec3f {
     let r1 = rng.gen_range(0.0..1.0);
     let r2 = rng.gen_range(0.0..1.0);
     let phi = 2.0 * PI * r1;
-    let theta = f32::acos(r2);
+    let theta = f64::acos(r2);
     Vec3f::normalize(from_spherical(theta, phi))
 }
 
@@ -105,20 +105,20 @@ pub fn cosine_weighted_hemisphere() -> Vec3f {
     let r1 = rng.gen_range(0.0..1.0);
     let r2 = rng.gen_range(0.0..1.0);
     let phi = 2.0 * PI * r1;
-    let theta = f32::acos(f32::sqrt(r2));
+    let theta = f64::acos(f64::sqrt(r2));
     Vec3f::normalize(from_spherical(theta, phi))
 }
 
 // https://agraphicsguy.wordpress.com/2015/11/01/sampling-microfacet-brdf/
 // https://computergraphics.stackexchange.com/questions/4979/what-is-importance-sampling
 // https://schuttejoe.github.io/post/ggximportancesamplingpart1/
-pub fn ggx_hemisphere(roughness: f32) -> Vec3f {
+pub fn ggx_hemisphere(roughness: f64) -> Vec3f {
     let mut rng = rand::thread_rng();
     let r1 = rng.gen_range(0.0..1.0);
     let r2 = rng.gen_range(0.0..1.0);
     let phi = 2.0 * PI * r1;
     let a2 = roughness * roughness;
-    let theta = f32::acos(f32::sqrt(a2 / (r2 * (a2 - 1.0) + 1.0)));
+    let theta = f64::acos(f64::sqrt(a2 / (r2 * (a2 - 1.0) + 1.0)));
     Vec3f::normalize(from_spherical(theta, phi))
 }
 
@@ -146,11 +146,11 @@ pub fn from_hex(color: u32) -> Vec3f {
     let r = (color & 0xff0000) >> 16;
     let g = (color & 0x00ff00) >> 8;
     let b = color & 0x0000ff;
-    Vec3f::new(r as f32, g as f32, b as f32) / (u8::MAX as f32)
+    Vec3f::new(r as f64, g as f64, b as f64) / (u8::MAX as f64)
 }
 
 pub fn to_image(framebuffer: Vec<Vec3f>, width: u32, height: u32) -> RgbImage {
-    let scale = u8::MAX as f32;
+    let scale = u8::MAX as f64;
 
     let buffer: Vec<u8> = framebuffer
         .iter()
@@ -179,8 +179,8 @@ mod test {
             let vec = sample_hemisphere();
             let sample = (vec + 1.0) / 2.0;
 
-            let w = width as f32;
-            let h = height as f32;
+            let w = width as f64;
+            let h = height as f64;
 
             let x = (sample.x.clamp(0.0, 0.999) * w) as usize;
             let y = (sample.y.clamp(0.0, 0.999) * h) as usize;
